refactor: #1424 Add parent class to contribution events
- Add `ContributionEvent` as `@MappedSuperclass`
