+++ src/main/java/ai/elimu/model/contributor/AudioContributionEvent.java
@@ -1,99 +1,21 @@
 package ai.elimu.model.contributor;
 
-import ai.elimu.model.BaseEntity;
 import ai.elimu.model.content.multimedia.Audio;
-import ai.elimu.model.v2.enums.Platform;
-import java.util.Calendar;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 @Entity
-public class AudioContributionEvent extends BaseEntity {
-    
-    @NotNull
-    @ManyToOne
-    private Contributor contributor;
+public class AudioContributionEvent extends ContributionEvent {
     
     @NotNull
     @ManyToOne
     private Audio audio;
-    
-    @NotNull
-    private Integer revisionNumber;
-
-    @Column(length = 1000)
-    private String comment;
-    
-    @NotNull
-    @Temporal(TemporalType.TIMESTAMP)
-    private Calendar time;
-    
-    /**
-     * The time passed during the creation/editing of the {@link #audio}.
-     */
-    @NotNull
-    private Long timeSpentMs;
-    
-    /**
-     * The platform that the {@link Contributor} used.
-     */
-    @NotNull
-    @Enumerated(EnumType.STRING)
-    private Platform platform;
-
-    public String getComment() {
-        return comment;
-    }
-
-    public Contributor getContributor() {
-        return contributor;
-    }
-
-    public Calendar getTime() {
-        return time;
-    }
 
     public Audio getAudio() {
         return audio;
     }
-    
-    public Integer getRevisionNumber() {
-        return revisionNumber;
-    }
-
-    public void setRevisionNumber(Integer revisionNumber) {
-        this.revisionNumber = revisionNumber;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public void setContributor(Contributor contributor) {
-        this.contributor = contributor;
-    }
-
-    public void setTime(Calendar time) {
-        this.time = time;
-    }
 
     public void setAudio(Audio audio) {
         this.audio = audio;
-    }
-
-    public Long getTimeSpentMs() {
-        return timeSpentMs;
-    }
-
-    public void setTimeSpentMs(Long timeSpentMs) {
-        this.timeSpentMs = timeSpentMs;
-    }
-    
-    public Platform getPlatform() {
-        return platform;
-    }
-    
-    public void setPlatform(Platform platform) {
-        this.platform = platform;
     }
 }
