+++ src/main/java/ai/elimu/model/contributor/NumberContributionEvent.java
@@ -1,83 +1,21 @@
 package ai.elimu.model.contributor;
 
-import ai.elimu.model.BaseEntity;
 import ai.elimu.model.content.Number;
-import java.util.Calendar;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 @Entity
-public class NumberContributionEvent extends BaseEntity {
-    
-    @NotNull
-    @ManyToOne
-    private Contributor contributor;
+public class NumberContributionEvent extends ContributionEvent {
     
     @NotNull
     @ManyToOne
     private Number number;
-    
-    @NotNull
-    private Integer revisionNumber;
-
-    @Column(length = 1000)
-    private String comment;
-    
-    @NotNull
-    @Temporal(TemporalType.TIMESTAMP)
-    private Calendar time;
-    
-    /**
-     * The time passed during the creation/editing of the {@link #number}.
-     */
-    @NotNull
-    private Long timeSpentMs;
-
-    public String getComment() {
-        return comment;
-    }
-
-    public Contributor getContributor() {
-        return contributor;
-    }
-
-    public Calendar getTime() {
-        return time;
-    }
 
     public Number getNumber() {
         return number;
     }
-    
-    public Integer getRevisionNumber() {
-        return revisionNumber;
-    }
-
-    public void setRevisionNumber(Integer revisionNumber) {
-        this.revisionNumber = revisionNumber;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public void setContributor(Contributor contributor) {
-        this.contributor = contributor;
-    }
-
-    public void setTime(Calendar time) {
-        this.time = time;
-    }
 
     public void setNumber(Number number) {
         this.number = number;
-    }
-
-    public Long getTimeSpentMs() {
-        return timeSpentMs;
-    }
-
-    public void setTimeSpentMs(Long timeSpentMs) {
-        this.timeSpentMs = timeSpentMs;
     }
 }
