+++ src/main/resources/META-INF/jpa-schema-export.sql
@@ -217,8 +217,8 @@
         revisionNumber integer,
         time timestamp,
         timeSpentMs bigint,
-        audio_id bigint,
         contributor_id bigint,
+        audio_id bigint,
         primary key (id)
     );
 
@@ -399,6 +399,7 @@
     create table NumberContributionEvent (
        id bigint generated by default as identity (start with 1),
         comment varchar(1000),
+        platform varchar(255),
         revisionNumber integer,
         time timestamp,
         timeSpentMs bigint,
@@ -686,14 +687,14 @@
        references Audio;
 
     alter table AudioContributionEvent 
-       add constraint FKspea1r50sj31ovaw0cmsrdd1t 
-       foreign key (audio_id) 
-       references Audio;
-
-    alter table AudioContributionEvent 
        add constraint FKk5x3wa0d4qp54r94o7tky3mrt 
        foreign key (contributor_id) 
        references Contributor;
+
+    alter table AudioContributionEvent 
+       add constraint FKspea1r50sj31ovaw0cmsrdd1t 
+       foreign key (audio_id) 
+       references Audio;
 
     alter table AudioPeerReviewEvent 
        add constraint FK9750pa9ak23p1y0lwft00vkch
